move cli resource views to ol/types
