+++ json-rpc/types/Cargo.toml
@@ -23,3 +23,5 @@ libra-workspace-hack = { path = "../../common/workspace-hack", version = "0.1.0"
 move-core-types = { path = "../../language/move-core/types", version = "0.1.0" }
 transaction-builder = { path = "../../language/transaction-builder", version = "0.1.0"}
 move-explain = { path = "../../language/tools/move-explain", version = "0.1.0" }
+# //////// 0L ////////
+ol-types = { path = "../../ol/types" }
