+++ ol/types/Cargo.toml
@@ -27,4 +27,6 @@ url = { version = "^2.1", features = ["serde"]}
 dialoguer = "0.8.0"
 once_cell = "1.7.2"
 keygen = { path = "../keygen" }
-
+move-core-types = { path = "../../language/move-core/types", version = "0.1.0" }
+sha2 = "0.9.1"
+serde-hex = "0.1.0"
