+++ ol/types/src/miner_state.rs
@@ -1,6 +1,6 @@
-// add by Ping
+//! miner state view for cli
 
-use crate::{
+use libra_types::{
     access_path::AccessPath,
     account_config::constants:: CORE_CODE_ADDRESS,
 };
@@ -15,13 +15,19 @@ use move_core_types::account_address::AccountAddress;
 /// Struct that represents a CurrencyInfo resource
 #[derive(Debug, Serialize, Deserialize)]
 pub struct MinerStateResource {
+    ///
     pub previous_proof_hash: Vec<u8>,
-    // pub reported_tower_height: u64,
-    pub verified_tower_height: u64, // user's latest verified_tower_height
+    /// user's latest verified_tower_height
+    pub verified_tower_height: u64, 
+    ///
     pub latest_epoch_mining: u64,
+    ///
     pub count_proofs_in_epoch: u64,
+    ///
     pub epochs_validating_and_mining: u64,
+    ///
     pub contiguous_epochs_validating_and_mining: u64,
+    ///
     pub epochs_since_last_account_creation: u64,
 }
 
@@ -31,7 +37,7 @@ impl MoveResource for MinerStateResource {
 }
 
 impl MinerStateResource {
-
+    ///
     pub fn struct_tag() -> StructTag {
         StructTag {
             address: CORE_CODE_ADDRESS,
@@ -40,7 +46,7 @@ impl MinerStateResource {
             type_params: vec![],
         }
     }
-
+    ///
     pub fn access_path(account: AccountAddress) -> AccessPath {
         let resource_key = ResourceKey::new(
             account,
@@ -48,11 +54,12 @@ impl MinerStateResource {
         );
         AccessPath::resource_access_path(&resource_key)
     }
-
+    ///
     pub fn resource_path() -> Vec<u8> {
         AccessPath::resource_access_vec(&MinerStateResource::struct_tag())
     }
 
+    /// 
     pub fn try_from_bytes(bytes: &[u8]) -> Result<Self> {
         lcs::from_bytes(bytes).map_err(Into::into)
     }
