+++ types/src/account_config/resources/mod.rs
@@ -15,8 +15,8 @@ pub mod role_id;
 pub mod vasp;
 pub mod withdraw_capability;
 //////// 0L ////////
-pub mod miner_state;
-pub mod oracle_upgrade;
+// pub mod miner_state;
+// pub mod oracle_upgrade;
 
 pub use account::*;
 pub use balance::*;
